refactoring: make BaseEntity type is generics and other entities id are Long

+++ shoppingOnline/src/main/java/com/roninEngineerF02/shoppingOnline/repository/CartRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CartRepository  extends JpaRepository<Cart, Integer> {
-    Optional<Cart> findByUser_Id(Integer userId);
+public interface CartRepository  extends JpaRepository<Cart, Long> {
+    Optional<Cart> findByUser_Id(Long userId);
 }
